+++ examples/hello-world/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const sentence = require('../../src')
-
-console.log(sentence().get())