+++ examples/myman-insult-formula/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const Insult = require('./Insult.js')
-
-setInterval(() => console.log(new Insult().get()), 2000)