+++ src/defaults/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  options: require('./options.json'),
-  templates: require('./templates.json'),
-  vocabulary: require('./vocabulary.json'),
-};