+++ src/factory/defaults/index.ts
@@ -1,0 +1,7 @@
+import templates from './templates.json';
+import vocabulary from './vocabulary.json';
+
+export default {
+  templates,
+  vocabulary,
+};