+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 
+## [0.5.0](../../tree/v0.5.0) - 16.08.2020
+### Breaking :boom:
+* **Refactored** entire project to TypeScript
+* **Removed** default export, should now defer to importing `createSentence` for the default behaviour
+* **Removed** or **changed** all other exports
+### Features :tada:
+* **Added** class `Sentence` to exports
+* **Added** class `SentenceFactory` to exports
+
 ## [0.4.2](../../tree/v0.4.2) - 11.08.2020
-#### Documentation
+### Documentation
 * :nut_and_bolt: **Change** package keywords
 * :nut_and_bolt: **Change** package description
 * :nut_and_bolt: **Change** README
@@ -14,10 +23,10 @@
 * :bug: **Fix** `setOptions` so that it no longer resets unspecified options to previous defaults
 
 ## [0.4.0](../../tree/v0.4.0) - 26.07.2020
-#### Features
+### Features
 * :tada: **New** option `forceNewSentence`, which will force `Sentence.generate()` to avoid duplicate of previous sentence when possible
 * :tada: **New** method `restoreDefaults()` available at entry point
-#### Development
+### Development
 * :nut_and_bolt: **Add** jest for testing
 * :nut_and_bolt: **Add** eslint for linting
 * :nut_and_bolt: **Add** .npmignore